requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cssselect==1.4.0
cymem==2.0.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.25.2
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
newspaper3k==0.2.8
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
preshed==3.0.12
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rich==14.3.3
setuptools==82.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.3.10
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.2
